Removed all dependencies on native code, use lib NetCDF-Java instead. Bug that prevented me from using that library was fixed! Wrote code to use NetCDF-Java for reading BeamFormed data. Fixed Unit tests, all tests pass now. Updated scripte to remove native code stuff. Fixed package name for tests (no capitals).
Added notices.
Updated jacoco.
