+++ TODO.md
@@ -2,4 +2,4 @@ add to GUI: controls to choose station, polarization / stoke, slider + checkbox 
 - show amplitude graph per channel
 - show amplitude graph total
 - show SNR, min, max, avg, mean, median per channel and total
-- make seperate tabs? One for graph, one for stats, ...
+- make separate tabs? One for graph, one for stats, ...
