+++ pulse-profile
@@ -1,10 +1,3 @@
 #! /bin/bash
 
-LIBRARY="dist/linux_x86"
-
-if [ `/bin/uname -i` = "x86_64" ]
-then
-    LIBRARY="dist/linux_x86_64"
-fi
-
-LD_LIBRARY_PATH=$LIBRARY:$LD_LIBRARY_PATH java -Xmx2048M -classpath "dist/*:." nl.esciencecenter.eastroviz.PulseProfile $*
+java -Xmx2048M -classpath "dist/*:." nl.esciencecenter.eastroviz.PulseProfile $*
