+++ src/nl/esciencecenter/eastroviz/dataformats/beamformed/BeamFormedDataReader.java
@@ -7,22 +7,17 @@ import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.nio.FloatBuffer;
 import java.nio.channels.FileChannel;
-import java.util.List;
-
-import ncsa.hdf.object.Attribute;
-import ncsa.hdf.object.DataFormat;
-import ncsa.hdf.object.FileFormat;
-import ncsa.hdf.object.HObject;
-import ncsa.hdf.object.h5.H5Group;
-import ncsa.hdf.object.h5.H5ScalarDS;
-import nl.esciencecenter.eastroviz.Viz;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import nl.esciencecenter.eastroviz.Viz;
+import ucar.nc2.Group;
+import ucar.nc2.NetcdfFile;
+import ucar.nc2.Variable;
+
 public class BeamFormedDataReader {
     private static final Logger LOGGER = LoggerFactory.getLogger(BeamFormedDataReader.class);
-    private FileFormat fileFormat;
     private String rawFileName;
     private String hdf5FileName;
 
@@ -54,115 +49,82 @@ public class BeamFormedDataReader {
         LOGGER.info("hdf5 file = " + hdf5FileName + ", raw file = " + rawFileName);
     }
 
-    public BeamFormedMetaData readMetaData() {
-        // retrieve an instance of H5File
-        final FileFormat fileFormat1 = FileFormat.getFileFormat(FileFormat.FILE_TYPE_HDF5);
-        if (fileFormat1 == null) {
-            LOGGER.info("Cannot find HDF5 FileFormat.");
-            System.exit(1);
-        }
+    public BeamFormedMetaData readMetaData() throws IOException {
+        NetcdfFile ncfile = null;
+        ncfile = NetcdfFile.open(hdf5FileName);
 
-        try {
-            fileFormat = fileFormat1.createInstance(hdf5FileName, FileFormat.READ);
-            if (fileFormat == null) {
-                LOGGER.info("Cannot open HDF5 File.");
-                System.exit(1);
-            }
+        Group rootGroup = ncfile.getRootGroup();
+        
+        LOGGER.info("group = " + rootGroup.getNameAndAttributes());
 
-            // open the file and retrieve the file structure
-            @SuppressWarnings("unused")
-            final int fileID = fileFormat.open();
-        } catch (final Exception e) {
-            LOGGER.info("" + e);
-        }
-
-        final H5Group root = (H5Group) ((javax.swing.tree.DefaultMutableTreeNode) fileFormat.getRootNode()).getUserObject();
-        LOGGER.info("root object: " + root + " type is: " + root.getClass() + " full name = " + root.getFullName());
-        final java.util.List<HObject> rootMemberList = root.getMemberList();
-        for (int i = 0; i < rootMemberList.size(); i++) {
-            LOGGER.info("root member " + i + " = " + rootMemberList.get(i));
-        }
-        printAttributes(root);
-
-        // Sanity check data type
-        if (!getAttribute(root, "FILETYPE").getPrimitiveStringVal().equals("bf")) {
+        if(!rootGroup.findAttribute("FILETYPE").getStringValue().equals("bf")) {
             throw new RuntimeException("file type is not bf!");
         }
+        
+        m.nrStations = rootGroup.findAttribute("OBSERVATION_NOF_STATIONS").getNumericValue().intValue();
+        LOGGER.info("nrStations: " + m.nrStations);
 
-        m.nrStations = getAttribute(root, "OBSERVATION_NOF_STATIONS").getPrimitiveIntVal();
-        LOGGER.info("nrStations = " + m.nrStations);
-
-        m.totalIntegrationTime = getAttribute(root, "TOTAL_INTEGRATION_TIME").getPrimitiveDoubleVal();
+        m.totalIntegrationTime = rootGroup.findAttribute("TOTAL_INTEGRATION_TIME").getNumericValue().doubleValue();
         LOGGER.info("total integration time = " + m.totalIntegrationTime);
 
-        m.bitsPerSample = getAttribute(root, "OBSERVATION_NOF_BITS_PER_SAMPLE").getPrimitiveIntVal();
+        m.bitsPerSample = rootGroup.findAttribute("OBSERVATION_NOF_BITS_PER_SAMPLE").getNumericValue().intValue();
         LOGGER.info("bits per sample = " + m.bitsPerSample);
 
-        m.clockFrequency = getAttribute(root, "CLOCK_FREQUENCY").getPrimitiveDoubleVal();
+        m.clockFrequency = rootGroup.findAttribute("CLOCK_FREQUENCY").getNumericValue().doubleValue();
         LOGGER.info("clock frequency = " + m.clockFrequency + " MHz");
 
-        m.minFrequency = getAttribute(root, "OBSERVATION_FREQUENCY_MIN").getPrimitiveDoubleVal();
+        m.minFrequency = rootGroup.findAttribute("OBSERVATION_FREQUENCY_MIN").getNumericValue().doubleValue();
         LOGGER.info("min frequency = " + m.minFrequency);
 
-        m.maxFrequency = getAttribute(root, "OBSERVATION_FREQUENCY_MAX").getPrimitiveDoubleVal();
+        m.maxFrequency = rootGroup.findAttribute("OBSERVATION_FREQUENCY_MAX").getNumericValue().doubleValue();
         LOGGER.info("max frequency = " + m.maxFrequency);
 
-        // first member is pointing 0
-        final H5Group pointing0 = (H5Group) rootMemberList.get(0);
-        LOGGER.info("-- start pointing 0 --");
-        printAttributes(pointing0);
-        LOGGER.info("-- end pointing 0 --");
-
-        m.nrBeams = getAttribute(pointing0, "NOF_BEAMS").getPrimitiveIntVal();
+        // first member is log, 2nd member is pointing 0
+        Group pointing0 = rootGroup.getGroups().get(1);
+        LOGGER.info("pointing0 = " + pointing0);
+        LOGGER.info("attributes: " + pointing0.getAttributes());
+        
+        m.nrBeams = pointing0.findAttribute("NOF_BEAMS").getNumericValue().intValue();
         LOGGER.info("nrBeams = " + m.nrBeams);
 
         // first member of the pointing is beam 0
-        final java.util.List<HObject> pointing0MemberList = pointing0.getMemberList();
-        for (int i = 0; i < pointing0MemberList.size(); i++) {
-            LOGGER.info("pointing 0 member " + i + " = " + pointing0MemberList.get(i));
-        }
+        LOGGER.info( "pointing0 groups: " + pointing0.getGroups());
 
-        final H5Group beam0 = (H5Group) pointing0MemberList.get(0);
-        printAttributes(beam0);
+        Group beam0 = pointing0.getGroups().get(1);
+        LOGGER.info( "beam0 attributes: " + beam0.getAttributes());
 
-        final java.util.List<HObject> beam0MemberList = beam0.getMemberList();
-        for (int i = 0; i < beam0MemberList.size(); i++) {
-            LOGGER.info("beam 0 member " + i + " = " + beam0MemberList.get(i));
-        }
-
-        m.totalNrSamples = getAttribute(beam0, "NOF_SAMPLES").getPrimitiveIntVal();
+        m.totalNrSamples = beam0.findAttribute("NOF_SAMPLES").getNumericValue().intValue();
         LOGGER.info("totalNrSamples = " + m.totalNrSamples);
 
-        m.nrChannels = getAttribute(beam0, "CHANNELS_PER_SUBBAND").getPrimitiveIntVal();
+        m.nrChannels = beam0.findAttribute("CHANNELS_PER_SUBBAND").getNumericValue().intValue();
         LOGGER.info("nrChannels = " + m.nrChannels);
 
-        m.nrStokes = getAttribute(beam0, "NOF_STOKES").getPrimitiveIntVal();
+        m.nrStokes = beam0.findAttribute("NOF_STOKES").getNumericValue().intValue();
         LOGGER.info("nrStokes = " + m.nrStokes);
         if (m.nrStokes != 1) {
             throw new UnsupportedOperationException("only 1 stoke supported for now.");
         }
 
-        m.subbandWidth = getAttribute(beam0, "SUBBAND_WIDTH").getPrimitiveDoubleVal() / 1000000.0f;
-        m.channelWidth = getAttribute(beam0, "CHANNEL_WIDTH").getPrimitiveDoubleVal() / 1000000.0f;
-        m.beamCenterFrequency = getAttribute(beam0, "BEAM_FREQUENCY_CENTER").getPrimitiveDoubleVal();
+        m.subbandWidth = beam0.findAttribute("SUBBAND_WIDTH").getNumericValue().doubleValue() / 1000000.0;
+        m.channelWidth = beam0.findAttribute("CHANNEL_WIDTH").getNumericValue().doubleValue() / 1000000.0;
+        m.beamCenterFrequency = beam0.findAttribute("BEAM_FREQUENCY_CENTER").getNumericValue().doubleValue();
         LOGGER.info("subbandWidth = " + m.subbandWidth + ", channelWidth = " + m.channelWidth + ", beam center frequency = "
                 + m.beamCenterFrequency);
 
-        final String stokesComponents = getAttribute(beam0, "STOKES_COMPONENTS").getPrimitiveStringVal();
-        LOGGER.info("Stokes components = " + stokesComponents);
+        final String stokesComponents = beam0.findAttribute("STOKES_COMPONENTS").getStringValue();
+        LOGGER.info("Stokes components = Stokes" + stokesComponents);
         if (!stokesComponents.equals("I")) {
             throw new UnsupportedOperationException("only stokes I supported for now.");
         }
 
         // Third member of the beam is the first stoke. First is "COORDINATES", 2nd = PROCESS_HISTORY
-        final H5ScalarDS stokes0 = (H5ScalarDS) beam0MemberList.get(2);
-        printAttributes(stokes0);
+        LOGGER.info("beam0 groups: " + beam0.getGroups());
+        LOGGER.info("beam0 vars: " + beam0.getVariables());
 
-        final int[] nrChannelsArray = getAttribute(stokes0, "NOF_CHANNELS").get1DIntArrayVal();
-        m.nrSubbands = nrChannelsArray.length;
+        Variable v = beam0.getVariables().get(0);
+        m.nrSubbands = v.findAttribute("NOF_SUBBANDS").getNumericValue().intValue();
         LOGGER.info("nrSubbands = " + m.nrSubbands);
-
-
+       
         if (zoomFactor >= 1) {
             m.nrTimes = (int) (m.totalIntegrationTime * zoomFactor);
             m.nrSamplesPerTimeStep = (int) (m.totalNrSamples / (m.totalIntegrationTime * zoomFactor));
@@ -176,14 +138,16 @@ public class BeamFormedDataReader {
         }
         LOGGER.info("nrSeconds = " + m.nrTimes + ", nrSamplesPerTimeStep = " + m.nrSamplesPerTimeStep);
 
+        ncfile.close();
+        
         return m;
     }
 
-    public BeamFormedData read() {
+    public BeamFormedData read() throws IOException {
         return read(null, null);
     }
 
-    public BeamFormedData read(BeamFormedMetaData m, BeamFormedSampleHandler handler) {
+    public BeamFormedData read(BeamFormedMetaData m, BeamFormedSampleHandler handler) throws IOException {
         if (m == null) {
             m = readMetaData();
         }
@@ -263,51 +227,8 @@ public class BeamFormedDataReader {
                     // ignore
                 }
             }
-
-            try {
-                fileFormat.close();
-                fileFormat = null;
-            } catch (final Exception e) {
-                e.printStackTrace();
-                System.exit(1);
             }
-        }
 
         return new BeamFormedData(fileName, maxSequenceNr, maxSubbands, zoomFactor, samples, initialFlagged, m);
-    }
-
-    @SuppressWarnings("unchecked")
-    private Hdf5Attribute getAttribute(final DataFormat node, final String name) {
-        List<Attribute> l = null;
-
-        try {
-            l = node.getMetadata();
-        } catch (final Exception e) {
-            throw new RuntimeException(e);
-        }
-
-        for (int i = 0; i < l.size(); i++) {
-            final Attribute a = l.get(i);
-            if (a.getName().equals(name)) {
-                return new Hdf5Attribute(a);
-            }
-        }
-        return null;
-    }
-
-    private void printAttributes(final DataFormat node) {
-        try {
-            @SuppressWarnings("unchecked")
-            final List<Attribute> metaList = node.getMetadata();
-            for (int i = 0; i < metaList.size(); i++) {
-                final Attribute a = metaList.get(i);
-                final Hdf5Attribute a2 = new Hdf5Attribute(a);
-                LOGGER.info("node: " + node.toString() + ": meta " + i + " = " + a.getName() + ", type = "
-                        + a.getType().getDatatypeDescription() + ", nrDims = " + a.getRank() + ", size = "
-                        + a.getType().getDatatypeSize() + ", val = " + a2.getValueString());
-            }
-        } catch (final Exception e1) {
-            e1.printStackTrace();
-        }
     }
 }
